Minor: Relocated pa_bt_test to features/pt_bt
